+++ CHANGELOG.md
@@ -9,5 +9,11 @@
 -->
 Please backup your data before updating to this version.
 
+## Additions
+- Extensions now can be trusted by repo
+
+## Changes
+- Extensions now required to have `repo.json`
+
 ## Other
 - Migrate to SQLDelight